test_get_enforced_params
